+++ jsondb/pom.xml
@@ -71,6 +71,12 @@
       <optional>true</optional>
     </dependency>
 
+    <!-- For test resource -->
+    <dependency>
+      <groupId>io.swagger</groupId>
+      <artifactId>swagger-annotations</artifactId>
+    </dependency>
+
     <!-- == Test ==========================================================================  -->
     <dependency>
       <groupId>com.h2database</groupId>
