+++ project-generator/pom.xml
@@ -26,8 +26,8 @@
   </parent>
   <modelVersion>4.0.0</modelVersion>
 
-  <artifactId>integration-to-maven-project-converter</artifactId>
-  <name>iPaaS REST :: Integration to Maven Project Converter</name>
+  <artifactId>project-generator</artifactId>
+  <name>iPaaS REST :: Project Generator</name>
 
   <repositories>
     <repository>
