+++ project-generator/src/main/java/com/redhat/ipaas/project/converter/ProjectGeneratorConfiguration.java
@@ -13,25 +13,18 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.redhat.ipaas.rest.v1.handler.jsondb;
+package com.redhat.ipaas.project.converter;
 
-import javax.ws.rs.Path;
-
-import com.redhat.ipaas.jsondb.JsonDB;
-import com.redhat.ipaas.jsondb.rest.JsonDBResource;
-import io.swagger.annotations.Api;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty;
+import com.redhat.ipaas.connector.catalog.ConnectorCatalog;
+import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
-import org.springframework.stereotype.Component;
 
-@Path("/jsondb")
-@Api(value = "jsondb")
-@Component
 @Configuration
-@ConditionalOnProperty(value = "endpoints.jsondb.enabled")
-public class JsonDBHandler extends JsonDBResource {
-    public JsonDBHandler(JsonDB jsondb) {
-        super(jsondb);
+public class ProjectGeneratorConfiguration {
+
+    @Bean
+    public ProjectGenerator projectConverter(ConnectorCatalog connectorCatalog) {
+        return new DefaultProjectGenerator(connectorCatalog);
     }
+
 }
