+++ project-generator/src/main/resources/com/redhat/ipaas/project/converter/templates/Application.java.mustache
