+++ project-generator/src/main/resources/com/redhat/ipaas/project/converter/templates/pom.xml.mustache
