+++ project-generator/src/test/resources/com/redhat/ipaas/project/converter/test-pull-push-application.yml
